// clients/company-supervisor/src/types/InternshipJournalsType.tsx
export type InternshipJournalsType = {
    id: number;
    internshipId: number;
    companySupervisorId: number;
    operationTime: number;
    journal: string;
    unitName: string;
    startingDate: string;
    endDate: string;
    createDate: string;
    updateDate: string;
    confirmation: boolean;
  };
  